restructuring memory api

+++ include/sld-stack.hpp
@@ -11,17 +11,17 @@ namespace sld {
 
     struct stack_t {
         addr start;
-        u32  size;
-        u32  position;
-        u32  save;
+        u64  size;
+        u64  position;
+        u64  save;
     };
 
     bool  stack_validate        (const stack_t& stack);
     bool  stack_reset           (stack_t& stack);
     bool  stack_reset_to_save   (stack_t& stack);
     bool  stack_save            (stack_t& stack);
-    void* stack_push            (stack_t& stack, const u32 size);
-    bool  stack_pull            (stack_t& stack, const u32 size);
+    void* stack_push            (stack_t& stack, const u64 size);
+    bool  stack_pull            (stack_t& stack, const u64 size);
 
     //-------------------------------------------------------------------
     // STACK LIST
